pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TriEthix"
version = "0.1.0"
description = "A TRIADIC BENCHMARK FOR ETHICAL ALIGNMENT IN FOUNDATION MODELS"
authors = [{name="Albert Barque-Duran"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "openai>=1.40.0",
  "anthropic>=0.34.0",
  "google-generativeai>=0.7.2",
  "requests>=2.32.3",
  "pandas>=2.2.2",
  "numpy>=1.26.4",
  "matplotlib>=3.8.4",
  "scipy>=1.13.0"
]

[project.scripts]
TriEthix = "triethix.cli:main"
